Wrote catch to prevent probOcc = 1 in data likelihood
The issue with infinite values in log(terrOcc) was caused by probOcc becoming = 1 early in the MCMC. This again was a consequence of ridiculously high values in totDens_std.

I wrote a catch now that ensures that probOcc, when used in the data likelihood, is < 1. Note that this is "treating symptoms", not the root cause. Suboptimal initial values for density are likely the root cause.

We can try to work with this "quick fix", but we need to screen our posteriors for ridiculously high density values.